feat(menu-list-nft): Allow the user to click list nft
It will navigate to nft details
